requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
filelock==3.15.4
fsspec==2024.6.1
idna==3.7
intel-openmp==2021.4.0   
Jinja2==3.1.4
MarkupSafe==2.1.5        
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.4.0
requests==2.32.3
sympy==1.13.0
tbb==2021.13.0
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2